Adds semi-deprecated design doc on V1 algorithm (#6)
This change exports the original design doc from within Google and
externalises it for a wider audience. It details the original approach
of relying on a Bazel BUILD graph to perform file-level splits, which
is no longer the preferred approach (except as a baseline to measure
algorithmic efficacy).

Of particular interest is some of the success criteria outlined to
measure the usability and efficacy of SplitBrain, taken from more of a
product point-of-view than purely an academic research project. There's
also some interesting commentary on different patterns that came up
during internal research (such as protocol buffers or configuration
changes being extremely common to split off). The simpler approach can
sometimes yield significant rewards for a minimum viable product.